Started the container rework
I made a new class called "TileContainer", which all containers that should contain tiles derives from. Deck, Hand, DiscardPile and PlayArea are all extended from TileContainer.

This way they all share common functions, saving a lot of confusion.

There is still work to be done with this, but it's stable for now.